refactor(ui): unify columns and group-by control composition
