+++ packages/ui/src/elements/ListGroupByButton/index.tsx
@@ -4,21 +4,18 @@ import type { ClientField, SanitizedCollectionConfig } from 'payload'
 import React, { useCallback } from 'react'
 
 import { useListQuery } from '../../providers/ListQuery/index.js'
-import { GroupByControl } from '../GroupByControl/index.js'
+import { GroupByButton } from '../GroupBy/index.js'
 
-export type ListGroupByControlProps = {
+export type ListGroupByButtonProps = {
   readonly collectionSlug: SanitizedCollectionConfig['slug']
   readonly fields: ClientField[]
 }
 
 /**
- * Connects the presentational {@link GroupByControl} to the list query state, so the
+ * Connects the presentational {@link GroupByButton} to the list query state, so the
  * control itself stays unaware of where its value is stored.
  */
-export const ListGroupByControl: React.FC<ListGroupByControlProps> = ({
-  collectionSlug,
-  fields,
-}) => {
+export const ListGroupByButton: React.FC<ListGroupByButtonProps> = ({ collectionSlug, fields }) => {
   const { query, refineListData } = useListQuery()
 
   const handleChange = useCallback(
@@ -29,7 +26,7 @@ export const ListGroupByControl: React.FC<ListGroupByControlProps> = ({
   )
 
   return (
-    <GroupByControl
+    <GroupByButton
       collectionSlug={collectionSlug}
       fields={fields}
       onChange={handleChange}
